requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
requests>=2.31.0
websocket-client>=1.6.0
